Restructure GitHub login into an Auth controller
Move the Socialite login flow into an invokable App\Http\Controllers\Auth\GithubController
that handles both the redirect and the callback on the same /login URL,
keeping the existing GitHub OAuth callback registration unchanged.
