+++ app/Http/Controllers/Auth/GithubController.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Auth;
 
 use App\Actions\GetOrCreateUser;
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Laravel\Socialite\Facades\Socialite;
@@ -12,8 +13,13 @@ use Laravel\Socialite\Two\InvalidStateException;
 use Symfony\Component\HttpFoundation\Response;
 use Throwable;
 
-class LoginController extends Controller
+class GithubController extends Controller
 {
+    /**
+     * GitHub redirects back to this same URL after authorization, so this
+     * entry point handles both phases: sending the user to GitHub, and
+     * processing the callback once GitHub returns with a `code`.
+     */
     public function __invoke(Request $request): Response
     {
         if ($request->user()) {
@@ -24,6 +30,14 @@ class LoginController extends Controller
             return Socialite::driver('github')->redirect();
         }
 
+        return $this->handleCallback($request);
+    }
+
+    /**
+     * Handle the callback from GitHub after authorization.
+     */
+    private function handleCallback(Request $request): Response
+    {
         try {
             $githubUser = Socialite::driver('github')->user();
         } catch (InvalidStateException) {
